% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqqpolt.R
\name{cs_qqplot1}
\alias{cs_qqplot1}
\title{chi-square distribution QQplot}
\usage{
cs_qqplot1(x, df)
}
\arguments{
\item{x}{the data}

\item{df}{the freedom of chi-square distribution}
}
\description{
This package is used for plot chis-square qqplot
}
\examples{
\dontrun{
a<-rchisq(100,3)
cs_qqplot1(a,3)
}
}
