% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{robenius_norm}
\alias{robenius_norm}
\title{frobenius norm}
\usage{
robenius_norm(x)
}
\arguments{
\item{x}{the matrix}
}
\value{
the norm of the matrix x
}
\description{
This package is used for calculate frobenius norm
}
\examples{
\dontrun{
robenius_norm(x)
}
}
